[WIP:feat] unattended setup script and other fixes

+++ _scripts/core/yazi-install.sh
@@ -23,7 +23,7 @@ case $CURRENT_OS_ID in
       cecho "yellow" "DRY-RUN: sudo pacman -S --noconfirm --needed ffmpegthumbnailer p7zip poppler imagemagick" 
     fi
     install_package "yazi" "yazi -V"
-  ;;
+    ;;
   debian|ubuntu)
     if [[ "$CURRENT_ARCH" == "aarch64" ]]; then
       source "$CDIR/rustup-install.sh"
@@ -37,14 +37,27 @@ case $CURRENT_OS_ID in
         cecho "yellow" "DRY-RUN: sudo apt install -y make gcc"
       fi
     else
-      . "$CDIR/homebrew-install.sh"
-      brew install ffmpegthumbnailer sevenzip poppler imagemagick
+      source "$CDIR/homebrew-install.sh"
+      if [ "$DRY_RUN" -ne "1" ]; then
+        brew install ffmpegthumbnailer sevenzip poppler imagemagick
+      else
+        cecho "yellow" "DRY-RUN: brew install ffmpegthumbnailer sevenzip poppler imagemagick"
+      fi
       install_package "yazi" "yazi -V" "brew install yazi"
     fi
-  ;;
+    ;;
+  fedora|redhat|centos|almalinux)
+    source "$CDIR/homebrew-install.sh"
+    if [ "$DRY_RUN" -ne "1" ]; then
+      brew install ffmpegthumbnailer sevenzip poppler imagemagick
+    else
+      cecho "yellow" "DRY-RUN: brew install ffmpegthumbnailer sevenzip poppler imagemagick"
+    fi
+    install_package "yazi" "yazi -V" "brew install yazi"
+    ;;
   *)
     cecho "red" "ERROR: Unsupported OS: $CURRENT_OS_ID!"
     exit 1
-  ;;
+    ;;
 esac
